[国标级联] 优化catalog推送代码

+++ src/main/java/com/genersoft/iot/vmp/gb28181/service/impl/PlatformChannelServiceImpl.java
@@ -215,6 +215,10 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
 
     @Transactional
     public int addChannelList(Integer platformId, List<CommonGBChannel> channelList) {
+        Platform platform = platformMapper.query(platformId);
+        if (platform == null) {
+            return 0;
+        }
         int result = platformChannelMapper.addChannels(platformId, channelList);
         if (result > 0) {
             // 查询通道相关的行政区划信息是否共享，如果没共享就添加
@@ -244,7 +248,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platformId, channelList, CatalogEvent.ADD);
+                eventPublisher.catalogEventPublish(platform, channelList, CatalogEvent.ADD);
             } catch (Exception e) {
                 log.warn("[关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -254,6 +258,11 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
 
     @Override
     public int removeAllChannel(Integer platformId) {
+        Platform platform = platformMapper.query(platformId);
+        if (platform == null) {
+            return 0;
+        }
+
         List<CommonGBChannel> channelListShare = platformChannelMapper.queryShare(platformId,  null);
         Assert.notEmpty(channelListShare, "未共享任何通道");
         int result = platformChannelMapper.removeChannelsWithPlatform(platformId, channelListShare);
@@ -278,7 +287,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platformId, channelListShare, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelListShare, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除全部关联通道] 发送失败，数量：{}", channelListShare.size(), e);
             }
@@ -302,6 +311,10 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
 
     @Transactional
     public int removeChannelList(Integer platformId, List<CommonGBChannel> channelList) {
+        Platform platform = platformMapper.query(platformId);
+        if (platform == null) {
+            return 0;
+        }
         int result = platformChannelMapper.removeChannelsWithPlatform(platformId, channelList);
         if (result > 0) {
             // 查询通道相关的分组信息
@@ -324,7 +337,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platformId, channelList, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelList, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -425,11 +438,12 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
     @Override
     public void updateCustomChannel(PlatformChannel channel) {
         platformChannelMapper.updateCustomChannel(channel);
+        Platform platform = platformMapper.query(channel.getPlatformId());
         CommonGBChannel commonGBChannel = platformChannelMapper.queryShareChannel(channel.getPlatformId(), channel.getGbId());
         // 发送消息
         try {
             // 发送catalog
-            eventPublisher.catalogEventPublish(channel.getPlatformId(), commonGBChannel, CatalogEvent.UPDATE);
+            eventPublisher.catalogEventPublish(platform, commonGBChannel, CatalogEvent.UPDATE);
         } catch (Exception e) {
             log.warn("[自定义通道信息] 发送失败， 平台ID： {}， 通道： {}（{}）", channel.getPlatformId(),
                     channel.getGbName(), channel.getId(), e);
@@ -468,7 +482,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -506,7 +520,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -537,7 +551,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
                 // 发送消息
                 try {
                     // 发送catalog
-                    eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.ADD);
+                    eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.ADD);
                 } catch (Exception e) {
                     log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
                 }
@@ -567,7 +581,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
                 // 发送消息
                 try {
                     // 发送catalog
-                    eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.ADD);
+                    eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.ADD);
                 } catch (Exception e) {
                     log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
                 }
